requirements.txt
requests
python-dotenv
google-generativeai
matplotlib
pandas
scikit-learn
fastapi
uvicorn